smtp_proxy_server/requirements.txt
aiosmtpd==1.4.5
pydantic-settings==2.2.1
requests==2.31.0
twisted==24.3.0
